+++ GetIT/app/src/main/java/com/getit/getit/ui/main/MainActivity.kt
@@ -1,7 +1,6 @@
 package com.getit.getit.ui.main
 
 import android.content.Intent
-import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.Toast
@@ -12,6 +11,7 @@ import androidx.navigation.ui.setupWithNavController
 import com.getit.getit.R
 import com.getit.getit.databinding.ActivityMainBinding
 import com.getit.getit.ui.BaseActivity
+import com.getit.getit.ui.main.searchproduct.SearchActivity
 
 
 class MainActivity: BaseActivity<ActivityMainBinding>(ActivityMainBinding::inflate) {
